1. save version to settings and check if this version on server then set as default choice
2. setup schedule button event
3. enable scheduling in backend
4. removed wx python and associated UI
5. add check that at least one day is selected
6. add callback for install once
7. small fixes. Test is finished successfully
8. fix in uninstallation, check for None
9. added progress stage

Related work items: #15

+++ downloader_backend.py
@@ -12,7 +12,6 @@ from collections import namedtuple, OrderedDict
 import requests
 
 import iss_templates
-import set_log
 
 artifactory_dict = OrderedDict([
     ('Austin', r'http://ausatsrv01.ansys.com:8080/artifactory'),
@@ -66,7 +65,7 @@ class Downloader:
         self.product_version = None
         self.setup_exe = None
 
-        set_log.set_logger()
+        set_logger()
 
         settings_path = self.parse_args()
         with open(settings_path, "r") as file:
@@ -192,8 +191,19 @@ class Downloader:
 
             self.zip_file = os.path.join(self.settings.download_path, f"temp_build_{self.settings.version}.zip")
             logging.info(f"Start download file from {url} to {self.zip_file}")
-            with open(self.zip_file, 'wb') as f:
-                shutil.copyfileobj(url_request.raw, f)
+
+            try:
+                file_size = int(url_request.headers['Content-Length'])
+            except TypeError:
+                file_size = 7e+9
+
+            percent = 0
+            with open(self.zip_file, 'wb') as zip_file:
+                for chunk in url_request.iter_content(chunk_size=int(file_size/100)):
+                    if chunk:
+                        logging.info(f"Download progress: {min(100, percent)}%")
+                        percent += 1
+                        zip_file.write(chunk)
 
             logging.info(f"File is downloaded to: {self.zip_file}")
 
@@ -219,6 +229,8 @@ class Downloader:
 
         if self.settings.delete_zip:
             self.clean_temp()
+
+        print("Installation finished!")
 
     def install_edt(self):
         """
@@ -278,7 +290,7 @@ class Downloader:
             logging.error("setup.exe does not exist")
             sys.exit(1)
 
-        if os.path.isfile(self.installed_product):
+        if self.installed_product and os.path.isfile(self.installed_product):
             uninstall_iss_file = os.path.join(self.target_unpack_dir, "uninstall.iss")
             uninstall_log_file = os.path.join(self.target_unpack_dir, "uninstall.log")
             with open(uninstall_iss_file, "w") as file:
@@ -447,6 +459,24 @@ class Downloader:
         else:
             logging.error("Please provide --path argument")
             sys.exit(1)
+
+
+def set_logger():
+    """
+    Function to setup logging output to stream and log file. Will be used by UI and backend
+    :return: None
+    """
+
+    settings_folder = os.path.join(os.environ["APPDATA"], "build_downloader")
+    logging_file = os.path.join(settings_folder, "downloader.log")
+
+    if not os.path.isdir(settings_folder):
+        os.mkdir(settings_folder)
+
+    # add logging to console and log file
+    logging.basicConfig(filename=logging_file, format='%(asctime)s (%(levelname)s) %(message)s', level=logging.DEBUG,
+                        datefmt='%d.%m.%Y %H:%M:%S')
+    logging.getLogger().addHandler(logging.StreamHandler())
 
 
 if __name__ == "__main__":
